backend/requirements.txt
openai-agents
cohere
qdrant-client
python-dotenv